requirements.txt
# =============================================================================
# AegisHarness — Python dependencies
# =============================================================================
# Install:  pip install -r requirements.txt
# =============================================================================

# LLM provider SDKs
openai>=1.30.0          # OpenAI + all OpenAI-compatible APIs (NVIDIA NIM, DeepSeek, Zhipu, etc.)
anthropic>=0.28.0       # Anthropic Claude API

# Config & environment
pyyaml>=6.0             # YAML config parsing (models_config.yaml)
python-dotenv>=1.0.0    # .env file loading

# Token accounting
tiktoken>=0.7.0         # Token counting for budget tracking

# Web API server
fastapi>=0.111.0        # REST + SSE API for the web management console
uvicorn[standard]>=0.30.0  # ASGI server (uvloop + httptools)

# =============================================================================
# v1.0.0 — Infrastructure & Persistence (M1)
# =============================================================================

# Database — async PostgreSQL + pgvector
sqlalchemy[asyncio]>=2.0.30     # ORM + async session management
asyncpg>=0.29.0                 # High-performance async PostgreSQL driver
alembic>=1.13.0                 # Schema migrations (alembic upgrade head)
# pgvector>=0.3.0               # pgvector Python client — uncomment in Week 4 (semantic search)

# =============================================================================
# v1.1.0 — Resilience & Parallelism (M2)
# =============================================================================

# Retry logic — exponential backoff + jitter for LLM API 429/5xx errors
tenacity>=8.2.0          # retry decorator (with_retry in retry_utils.py)

# =============================================================================
# v1.2.0 — Chat UI, MCP Manager, pgvector Semantic Search (M3 + M1 final)
# =============================================================================

# Vector embeddings — pgvector Python client (requires pgvector PG extension)
# pgvector>=0.3.0         # Uncomment when pgvector extension is available in DB
# Note: embeddings are stored as JSONB arrays; cosine similarity computed in Python.
# No hard dependency needed — VectorStore degrades gracefully without pgvector.

# Testing
pytest>=8.0.0           # Test runner
